% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnimgm.R
\docType{data}
\name{coc}
\alias{coc}
\title{A data set for Psychiatric Drug Treatment}
\format{A data frame with 869 rows and 10 variables}
\usage{
data(coc)
}
\description{
The variables are as follows:
\itemize{
  \item y. number of days
 \item time. in weeks
 \item sub. patients id
 \item group. treatment group with placebo (PBO) and desipramine (DMI)
 \item basey. baseline days
 \item g1. dummy variables for treatment group
 \item g2. dummy variables for treatment group
 \item time1. dummy variables for time
 \item time2. dummy variables for time
 \item grptime. group by time
}
}
\keyword{datasets}
