% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnigls.R
\name{expit}
\alias{expit}
\title{function for isnigls section expit: expit transformation}
\usage{
expit(u)
}
\arguments{
\item{u}{need more information}
}
\description{
function for isnigls section expit: expit transformation
}
