% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnigls.R
\name{expit2}
\alias{expit2}
\title{function for isnigls section  expit2: expit transformation extended}
\usage{
expit2(u, p)
}
\arguments{
\item{u}{need more information}

\item{p}{need more information}
}
\description{
function for isnigls section  expit2: expit transformation extended
}
