% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnigls.R
\name{fun.ar1mat}
\alias{fun.ar1mat}
\title{Internal Function for generate AR1 matrix and its first and second derivatives wrt D}
\usage{
fun.ar1mat(sigma, rho, obsi, transform, case = 1)
}
\arguments{
\item{sigma}{standard deviation}

\item{rho}{autoregressive  parameter}

\item{obsi}{dimision of the square matrix}

\item{transform}{logical indicating wether or not the parameter in D is transformed.}

\item{case}{1: return nobs x nobs AR1 matrix; 2: return a list for first derivatives of the compound symmetry matrix wrt sigma and rho;
3: eturn an array containing 2nd derivatives of the compound symmetry matrix wrt sigma and rho.}
}
\description{
Calculate the autoregressive matrix and its first and second derivatives wrt D.
}
