% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnigls.R
\name{fun.csmat}
\alias{fun.csmat}
\title{Internal Function for generate compound symmetry matrix and its first and second derivatives wrt D}
\usage{
fun.csmat(sigma, rho, obsi, transform, case = 1)
}
\arguments{
\item{sigma}{standard deviation}

\item{rho}{correlation coefficient}

\item{obsi}{dimension of the square matrix}

\item{transform}{logical indicating wether or not the parameter in D is transformed.}

\item{case}{1: return nobs x nobs CS matrix; 2: return a list for first derivatives of the compound symmetry matrix wrt sigma and rho;
3: eturn an array containing 2nd derivatives of the compound symmetry matrix wrt sigma and rho.}
}
\description{
Calculate the compound symmetry matrix and its first and second derivatives wrt D.
}
