% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnigls.R
\name{fun.generalmat}
\alias{fun.generalmat}
\title{Internal Function for generate a general symmetric matrix and its first and second derivatives wrt D}
\usage{
fun.generalmat(sigma, rho, obsi, transform, case = 1, case2 = NULL)
}
\arguments{
\item{sigma}{standard deviation}

\item{rho}{correlation coefficient}

\item{obsi}{dimension of the square matrix}

\item{transform}{logical indicating wether or not the parameter in D is transformed.}

\item{case}{1: return nobs x nobs general symmetric  matrix; 2: return a list for first derivatives of the general symmetric matrix wrt D;
3: return an array containing 2nd derivatives of the general symmetric matrix wrt D.}

\item{case2}{indicator for variance and covariance parameters.}
}
\description{
Calculate the general symmetric matrix and its first and second derivatives wrt D.
}
