% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnigls.R
\name{fun.gfit}
\alias{fun.gfit}
\title{Internal Function to fit the missing data model and obtain the predicted probabilities of being observed for all observations.}
\usage{
fun.gfit(data, gmodel)
}
\arguments{
\item{data}{the name of the dataset for fitting missing data mechansim model}

\item{gmodel}{a formula to specify a multinomial transitional missing data model.}
}
\description{
Fit missing data model and obtain predicted probabilities of being observed for all observations.
}
