% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isniglmmbin.R
\name{fun.ghqd}
\alias{fun.ghqd}
\title{Internal Function to integrate the function fun w.r.t. qpt using Gaussian-Hermite Quadrature}
\usage{
fun.ghqd(func, datai, ...)
}
\arguments{
\item{func}{function that contains the integration formula, the first argument must be qpt}

\item{datai}{list of data for subject i, must have below compnent: x: data frame representing covariates for beta.}

\item{...}{additional arguments to the function "func"}
}
\description{
Calculate subject-level quantities when the regression outcome is subject to missingness and 
follows linear mixed-effects  Models.
}
