% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isniglmmbin.R
\name{fun.subiq}
\alias{fun.subiq}
\title{Internal Function to calculate the various items (listed in calc) for subject i at quadrature point qpt.}
\usage{
fun.subiq(qpt, datai, par, calc = 1)
}
\arguments{
\item{qpt}{design matrix for random effects}

\item{datai}{list of data for subject i, must have below compnent: x: data frame representing covariates for beta.
y: vector representing the panel outcome.}

\item{par}{vector for the parameter beta and Sigma to be maximized}

\item{calc}{1: Calculate the conditional likelihood for subject i: f(Y_i|b_i);
2: Calculate the first derivative of item 1 above w.r.t beta and sigma;
4: return the three parts  of conditional mean's derivative w.r.t beta and sigma,}
}
\description{
Calculate the compound symmetry matrix and its first and second derivatives wrt the vari-cov matrix
}
