% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isniglm.R
\name{print.isniglm}
\alias{print.isniglm}
\title{Function to print  the isniglm object.}
\usage{
\method{print}{isniglm}(x, digits = max(3, getOption("digits") - 2), ...)
}
\arguments{
\item{x}{the isniglm object obtained from the isniglm function}

\item{digits}{the number of significant digits to use when printing}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function prints the model call, isni and c statistics from the isniglm object.
}
\description{
\code{print} method for class isniglm
}
