% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isniglmmbin.R
\name{print.isniglmm}
\alias{print.isniglmm}
\title{Function to print  the isniglmm object.}
\usage{
\method{print}{isniglmm}(x, digits = max(3, getOption("digits") - 2), ...)
}
\arguments{
\item{x}{the isniglmm object obtained from the isniglmm function}

\item{digits}{the number of significant digits to use when printing}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function print.isnilmm prints the model call,  isni and c statistics from the isniglmm object.
}
\description{
\code{print} method for class isniglmm
}
