% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnilmm.R
\name{print.isnilmm}
\alias{print.isnilmm}
\title{Function to print  the isnilmm object.}
\usage{
\method{print}{isnilmm}(x, digits = max(3, getOption("digits") - 2), ...)
}
\arguments{
\item{x}{the isnilmm object obtained from the isnilmm function}

\item{digits}{the number of significant digits to use when printing}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function print.isnilmm prints the model call,  isni and c statistics from the isnilmm object.
}
\description{
\code{print} method for class isnilmm
}
