% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnimgm.R
\name{print.isnimgm}
\alias{print.isnimgm}
\title{Function to print  the isnimgm object.}
\usage{
\method{print}{isnimgm}(x, digits = max(3, getOption("digits") - 2), ...)
}
\arguments{
\item{x}{the isnimgm object obtained from the isnimgm function}

\item{digits}{the number of significant digits to use when printing}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function print.isnimgm prints the model call, isni and c statistics from the isnimgm object.
}
\description{
\code{print} method for class isnimgm
}
