% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isniglm.R
\docType{data}
\name{r99}
\alias{r99}
\title{Dataset for a survey of sexual behavior}
\format{A data frame with 6136 rows and 5 variables}
\usage{
data(r99)
}
\description{
A dataset from a survey of sexual behavior
}
\details{
This data frame contains the following variables:
\itemize{
 \item  y: Response to the survey (1=yes, 0=no)
 \item gender: 0=male; 1=female
 \item fac: the student's faculty (1=medical/dental/veterinary, 0=other)
 \item genderbyfac: the product of gender and fac
 \item  g: an indicator of missing status (0=observed, 1=missing) for y
}
}
\keyword{datasets}
