% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isniglmmbin.R
\docType{data}
\name{skquit}
\alias{skquit}
\title{A randomzied trial data set for Smoking cessation}
\format{A data frame with 1861 rows and 6 variables}
\usage{
data(skquit)
}
\description{
The variables are as follows:
\itemize{
  \item id. subjects id
 \item  time. Time in weeks after baseline
 \item quit. 0: smoking; 1: quit smoking
 \item sub. patients id
 \item helmert1. Herlmert contrast 1 among 4 treatment groups.
 \item helmert2. Herlmert contrast 2 among 4 treatment groups.
 \item helmert3. Herlmert contrast 3 among 4 treatment groups.
}
}
\keyword{datasets}
