% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isniglm.R
\docType{data}
\name{sos}
\alias{sos}
\title{Dataset for a survey of sexual behavior}
\format{A data frame with 6136 rows and 3 variables}
\usage{
data(sos)
}
\description{
A dataset from a survey of sexual behavior
}
\details{
This data frame contains the following three factor variables:
\itemize{
 \item  sexact: response to the survey question "Have you ever had sexual intercouse" with two levels of 
                  \code{no} (ref level) and \code{yes}.
 \item gender: two levels of \code{male} (ref level) and female.
 \item faculty: the student's faculty with two levels of \code{mdv}=medical/dental/veterinary (ref level) and  \code{other}
               =all the other faculty categories.
}
}
\keyword{datasets}
