% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isniglm.R
\name{summary.isniglm}
\alias{summary.isniglm}
\title{Function to print out a summary of isniglm  object in a matrix form.}
\usage{
\method{summary}{isniglm}(object, ...)
}
\arguments{
\item{object}{the isniglm object obtained from the isniglm function}

\item{...}{additional arguements}
}
\description{
Function to print out a summary of isniglm  object in a matrix form.
}
