% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isniglmmbin.R
\name{summary.isniglmm}
\alias{summary.isniglmm}
\title{Function to print out a summary of isniglmm  object in a matrix form.}
\usage{
\method{summary}{isniglmm}(object, digits = max(3, getOption("digits") - 2), ...)
}
\arguments{
\item{object}{the isniglmm object obtained from the isniglmm function}

\item{digits}{the number of significant digits to use when printing}

\item{...}{additional arguments}
}
\description{
Function to print out a summary of isniglmm  object in a matrix form.
}
