% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnigls.R
\name{summary.isnigls}
\alias{summary.isnigls}
\title{Function to print out a summary of isnigls  object in a matrix form.}
\usage{
\method{summary}{isnigls}(object, ...)
}
\arguments{
\item{object}{the isniglm object obtained from the isnigls function}

\item{...}{additional arguements}
}
\description{
Function to print out a summary of isnigls  object in a matrix form.
}
