% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnilmm.R
\name{summary.isnilmm}
\alias{summary.isnilmm}
\title{Function to print out a summary of isnilmm  object in a matrix form.}
\usage{
\method{summary}{isnilmm}(object, ...)
}
\arguments{
\item{object}{the isnilmm object obtained from the isnigls function}

\item{...}{additional arguements}
}
\description{
Function to print out a summary of isnilmm  object in a matrix form.
}
