% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnimgm.R
\name{summary.isnimgm}
\alias{summary.isnimgm}
\title{Function to print out a summary of isnimgm  object in a matrix form.}
\usage{
\method{summary}{isnimgm}(object, digits = max(3, getOption("digits") - 2),
  ...)
}
\arguments{
\item{object}{the isnimgm object obtained from the isnimgm function}

\item{digits}{the number of significant digits to use when printing}

\item{...}{additional arguements}
}
\description{
Function to print out a summary of isnimgm  object in a matrix form.
}
