% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnimgm.R
\name{tmdm}
\alias{tmdm}
\title{Function to fit the transitional missing data model and obtain the predicted probabilities of being observed for all observations.}
\usage{
tmdm(formula, data, weights, subset)
}
\arguments{
\item{formula}{a formula to specify a multinomial transitional missing data model  in the form of g+gp ~Sterms.}

\item{data}{the name of the dataset for fitting missing data mechansim model}

\item{weights}{an optional vector of "prior weights" to be used in the fitting process for the missingness mechanism model.
Should be NULL or a numeric vector.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process for the outcome model and the missingness mechanism model.}
}
\description{
Fit missing data model and obtain predicted probabilities of being observed for all observations.
}
\examples{
qolefitg <- tmdm(g+gp~as.factor(time)+group+perf+sever,data=qolef)
}
