% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnigls.R
\docType{data}
\name{qolef}
\alias{qolef}
\title{A data set for Quality of Life Emontional Functioning outcome.}
\format{A data frame with 2860 rows and 13 variables}
\usage{
data(qolef)
}
\description{
The variables are as follows:
\itemize{
 \item id: patients id
 \item y: EF score
 \item time: time in months since randomization
 \item group: placebo (0) or flutamide (1)
 \item perf: baseline performance score
 \item sever: baseline disease severity
 \item t1: indicator variable for month 1
 \item t3: indicator variable for month 3
 \item t6: indicator variable for month 6
 \item yp: most recently observed  prior outcome
 \item basey: EF at baseline
 \item g: missingness status ("O"=observed, "D"=dropout, "I"=intermittent missingness)
 \item gp: missingness status in the prior visit ("O"=observed, "D"=dropout, "I"=intermittent missingness, "U"=undefined)
}
}
\keyword{datasets}
