% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isniglm.R
\name{summary.isniglm}
\alias{summary.isniglm}
\title{Function to print out a summary of isniglm  object in a matrix form.}
\usage{
\method{summary}{isniglm}(object, digits = max(3, getOption("digits") -
  2), ...)
}
\arguments{
\item{object}{the isniglm object obtained from the isniglm function}

\item{digits}{the number of significant digits to use when printing}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function summarizes the MAR coefficient estimates, standard errors, isni and c statistics from the isniglm object as a matrix form.
}
\description{
Function to print out a summary of isniglm  object in a matrix form.
}
