% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnilmm.R
\name{summary.isnilmm}
\alias{summary.isnilmm}
\title{Function to print out a summary of isnilmm  object in a matrix form.}
\usage{
\method{summary}{isnilmm}(object, digits = max(3, getOption("digits") - 2), ...)
}
\arguments{
\item{object}{the isnilmm object obtained from the isnilmm function}

\item{digits}{the number of significant digits to use when printing}

\item{...}{additional arguements}
}
\description{
Function to print out a summary of isnilmm  object in a matrix form.
}
